extract subsets distribution to separate component file
